+++ tests/test_runner.nim
@@ -1,3 +1,4 @@
 import unittest
 import test_template
 import test_context
+import test_glob